requirements.txt
openai~=1.109.0
python-dotenv~=1.1.1
numpy~=2.3.3
faiss-cpu~=1.12.0
sentence-transformers~=5.1.1
tqdm~=4.67.1
transformers~=4.56.2
aiogram~=3.22.0